requirements.txt
asgiref==3.7.2
Django==3.2.23
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==18.13.0
importlib-metadata==6.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
zipp==3.15.0
